+++ Material.Blazor.Website.Server/Startup.cs
@@ -26,7 +26,6 @@ namespace Material.BlazorWebsite.Server
             services.AddServerSideBlazor();
 
             services.AddMBServices(
-                animatedNavigationManagerServiceConfiguration: new(),
                 loggingServiceConfiguration: Utilities.GetDefaultLoggingServiceConfiguration(),
                 toastServiceConfiguration: Utilities.GetDefaultToastServiceConfiguration(),
                 snackbarServiceConfiguration: Utilities.GetDefaultSnackbarServiceConfiguration()